requirements.txt
# Web
fastapi
uvicorn[standard]
pydantic

# LangGraph / LangChain
langgraph
langchain-core
langchain-anthropic       # Claude Sonnet 4.6 (script_writer) + Haiku 4.5 (всі інші вузли)

# Queue
redis
celery

# Database
psycopg2-binary           # PostgreSQL
chromadb                  # Vector search (RAG)

# Indexer (file parsing — optional)
pypdf                     # PDF reader for book indexing
python-docx               # DOCX reader for book indexing

# Media (MVP 2+)
twelvelabs                # Video search
edge-tts                  # TTS (безкоштовний, Microsoft Azure voices)
mutagen                   # Audio duration detection

# Utils
python-dotenv
